feat(users): Add required constraint to referralId
This change updates the users model by adding a required constraint to referralId. Additionally, it modifies the user service to generate a referralId if none is provided during user registration. It also updates the signUpWithId and registeringMobile functions to improve their efficiency and fix a bug. Lastly, the verifyOTP function now correctly updates the user's verified status.