Add gitignore for build folder and IDEA project file, improve the verifyOtp and registeringMobile methods
This commit adds ".idea" folder to the gitignore and ignores the newline at the end of the build folder. The verifyOtp method now updates the verified field and returns if an existing user is being verified or a new user is being verified. The registeringMobile method now sets the existingUser field to false and has a message typo corrected.