Refactor UserService checkUserPhoneNumber and verifyOTP functions
The return type of checkUserPhoneNumber's Promise has been updated from void to any to return the value from the async function, and the verifyOTP function now returns a boolean instead of an ObjectId. The existingUser variable has been updated to use a ternary operator to directly return a boolean value.