Add Docker and CI/CD support to ReferAndEarn
The Dockerfile has been added with instructions to download the latest version of Node.js, install app dependencies, and configure an environment variable to run the server at port 8080. The run.sh script has been updated to build a Docker image of the app with a custom tag 'refer-and-earn-cabio'. Now users can pull & run this image anywhere, allowing for easier deployment of the ReferAndEarn application.